backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0